feat: Implement Azure Blob Storage sync for project folders
This commit introduces a new feature that allows users to synchronize a
local project document folder with an Azure Blob Storage container.

The Azure sync logic has been refactored into a separate
`azure_sync_service.py` file for better code organization.

Key features include:
- A new UI section in the Streamlit sidebar for configuring the sync settings.
- Users can specify a "Project Root Folder" and an "Azure Connection String".
- Configuration is saved to a `config.json` file within a `.fyiai` directory in the project root.
- "Push to Azure" and "Pull from Azure" buttons to perform the one-way sync.
- The Azure container is automatically named based on the project root folder (`fyiai-{project-name}`).
- The folder structure is preserved during the sync process.
- Added `azure-storage-blob` as a new dependency.
